requirements.txt
streamlit
streamlit-cropper
gdown
torch
torchvision
pillow
